Refactor in/out session command/parser files into correct package
