+++ src/main/java/seedu/system/logic/parser/outofsession/ListParticipationCommandParser.java
@@ -1,4 +1,4 @@
-package seedu.system.logic.parser.insession;
+package seedu.system.logic.parser.outofsession;
 
 import static seedu.system.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.system.logic.parser.CliSyntax.PREFIX_COMP;
