+++ src/main/java/seedu/system/model/session/Session.java
@@ -57,10 +57,8 @@ public class Session {
      * @param participations the list of participations in the competition session
      * @throws OngoingSessionException if there is already an ongoing session
      */
-    public void start(
-        Competition competition,
-        ObservableList<Participation> participations
-    ) throws OngoingSessionException {
+    public void start(Competition competition, ObservableList<Participation> participations)
+            throws OngoingSessionException {
         if (isOngoing()) {
             throw new OngoingSessionException();
         }
